Refactor: Enhance security, improve UI/UX, and fix JDK 24 compatibility
Backend:
- Replaced Lombok with standard Java to fix JDK 24 compatibility issues.
- Secured sensitive properties with environment variables.
- Implemented SecurityHeadersFilter for XSS and Clickjacking protection.
- Centralized CORS configuration.
- Sanitized error responses in GlobalExceptionHandler.
- Switched salary type to BigDecimal for precision.

Frontend:
- Updated department input to a smart dropdown with 'Other' option.
- Added email typo validation and correction suggestions.
- Implemented real-time XSS input sanitization and submission rate limiting.
- Improved accessibility with aria-labels and field IDs.
- Fixed form state reset bug when toggling Add/Edit modals.
- Configured environment variables for API base URL.

Infrastructure:
- Comprehensive SECURITY.md documentation and .env.example template.
- Hardened .gitignore.

+++ src/main/java/com/employee/management/dto/EmployeeResponseDto.java
@@ -1,29 +1,25 @@
 package com.employee.management.dto;
 
+import java.math.BigDecimal;
+
 public class EmployeeResponseDto {
 
     private Long id;
     private String name;
     private String department;
     private String email;
-    private Double salary;
-
- 
-
+    private BigDecimal salary;
 
     public EmployeeResponseDto() {
     }
 
-    public EmployeeResponseDto(Long id, String name, String department,
-                               String email, Double salary) {
+    public EmployeeResponseDto(Long id, String name, String department, String email, BigDecimal salary) {
         this.id = id;
         this.name = name;
         this.department = department;
         this.email = email;
         this.salary = salary;
-        }
-
-    // getters & setters
+    }
 
     public Long getId() {
         return id;
@@ -57,15 +53,11 @@ public class EmployeeResponseDto {
         this.email = email;
     }
 
-    public Double getSalary() {
+    public BigDecimal getSalary() {
         return salary;
     }
 
-    public void setSalary(Double salary) {
+    public void setSalary(BigDecimal salary) {
         this.salary = salary;
     }
-
-
-
-
 }
